Revert "Add use registered user HOC"
