+++ packages/commerce-sdk-react/src/hooks/ShopperCustomers/query.test.tsx
@@ -6,7 +6,7 @@
  */
 import React, {ReactElement} from 'react'
 import path from 'path'
-import {mockHttpResponses, renderWithProviders, withRegisteredUser} from '../../test-utils'
+import {mockHttpResponses, renderWithProviders} from '../../test-utils'
 import {
     useCustomer,
     useCustomerAddress,
@@ -25,13 +25,18 @@ const CUSTOMER_EMAIL = 'kobe@test.com'
 const ADDRESS_NAME = 'TestAddress'
 const LIST_ID = '987ae461a7c6c5fd17006fc774'
 
-const CustomerComponent = ({
-    customerId,
-    loginRegisteredUser
-}: {
-    customerId: string
-    loginRegisteredUser: any
-}): ReactElement => {
+const CustomerComponent = ({customerId}: {customerId: string}): ReactElement => {
+    const queryClient = useQueryClient()
+    const loginRegisteredUser = useShopperLoginHelper(ShopperLoginHelpers.LoginRegisteredUserB2C)
+
+    React.useEffect(() => {
+        queryClient.invalidateQueries([{entity: 'customer'}])
+        loginRegisteredUser.mutate({
+            username: 'kobe@test.com',
+            password: 'Test1234!'
+        })
+    }, [])
+
     const {data, isLoading, error} = useCustomer(
         {customerId},
         {
@@ -184,8 +189,7 @@ const tests = [
             {
                 name: 'returns data',
                 assertions: withMocks(async () => {
-                    const Component = withRegisteredUser(CustomerComponent)
-                    renderWithProviders(<Component customerId={CUSTOMER_ID} />)
+                    renderWithProviders(<CustomerComponent customerId={CUSTOMER_ID} />)
 
                     expect(screen.getByText('Loading...')).toBeInTheDocument()
                     await waitFor(() => screen.getByText(CUSTOMER_EMAIL))
@@ -195,8 +199,7 @@ const tests = [
             {
                 name: 'returns error',
                 assertions: withMocks(async () => {
-                    const Component = withRegisteredUser(CustomerComponent)
-                    renderWithProviders(<Component customerId={'WRONG_CUSTOMER_ID'} />)
+                    renderWithProviders(<CustomerComponent customerId={'WRONG_CUSTOMER_ID'} />)
 
                     expect(screen.getByText('Loading...')).toBeInTheDocument()
                     await waitFor(() => screen.getByText('error'))
